Read stock price from the same date as the reports
The balance sheets and income statements of companies are published on
quarterly or annual basis. Some statistics like Price to Earning ratio
(a.k.a P/E ratio) depend on both the stock price and the values
published in these reports.

It makes more sense to take the stock price of a company from the same
date as the report. This patch implements the following:
- Extraction of report release date from balance sheet/income statement
- Fetching of stock price given a date
- Calculation of different statistics based on the stock price at the
    date of the report

Signed-off-by: Shay Agroskin <agrosshay@gmail.com>