yfinance_info.py: cache stock price for date if market is closed
In get_stock_price_at_date() function the requested date might be a day
in which the market is closed. In this case the data returned from
yfinance won't include the requested date.

Make sure the date is cached in file nonetheless to avoid accessing the
internet every time for this case.