stocks_analyzer: Don't let exceptions in threads to unnoticed
The tickers are created using helper threads (each running a
create_ticker_worker() function instance).
If an exception occurs in one of these threads nothing is printed to the
screen which makes it very hard to debug.

This patch checks for exceptions for every line of code in the thread,
and makes sure to notify the user when something unexpected happens.

Signed-off-by: Shay Agroskin <agrosshay@gmail.com>